Add helper functions to better present information
Helper functions have been added to the User class in order to convert
from Map or List to String. This should help encapsulate information
presentation within the class and outside of the service and/or
controller.

All variables of type LocalDateTime have been converted to ZonedDateTime
in order to support explicit setting of a timezone. If not done, an
exception will be thrown when trying to format the time.

A DateTimeFormatter has been included as a static field in the
Transaction class to present time information in a simple manner while
also including the specified timezone (in this case, GMT+0).

A helper function has been added to the Transaction class to return the
properly formatted date and time information.