Fix bugs regarding logic in point add/deduct
As defined in the unit tests, a negative input for addPoints() will
result in no action taken and the transaction ignored.

A bug was found in the logic of deductPoints() when one prior
transaction was sufficient for the operation to complete. The bug was
reading information in the incorrect order, causing point values to be
inaccurate.